UserController -> RegisterController
moved registerNewUser method to RegisterController

removed UserController
